move navigation to subpackage
