+++ modules/web/src/com/haulmont/sample/petclinic/web/visit/visit/calendar/navigation/CalendarNavigators.java
@@ -1,10 +1,11 @@
-package com.haulmont.sample.petclinic.web.visit.visit.calendar;
+package com.haulmont.sample.petclinic.web.visit.visit.calendar.navigation;
 
 
 import com.haulmont.cuba.core.global.DatatypeFormatter;
 import com.haulmont.cuba.core.global.UserSessionSource;
 import com.haulmont.cuba.gui.components.Calendar;
 import com.haulmont.cuba.gui.components.DatePicker;
+import com.haulmont.sample.petclinic.web.visit.visit.calendar.CalendarMode;
 import org.springframework.stereotype.Component;
 
 import javax.inject.Inject;
